Initial implementation
This adds a basic implementation of our validator written in C++,
using the jwt-cpp library. Currently we only support OIDC providers that
use JWT (or JWT-ish in case of Azure) access tokens, as opaque tokens
will require provider-specific implementations in the future.

This is intended as a minimal proof-of-concept implementation, so it was
intentionally kept simple, no caching or any other advanced internal
features.

There's only on configuration option, the extension allows users to
select which field to use for identity mapping - it defaults to `sub`,
which should work most of the time, but it's not practical with some
vendors (Azure), in which case users can just select a different option.