Also check the "scope" attribute in the access token
The current code only checked the "scp" shorthad, but some providers
actually return the longer "scope" variant. The updated code accepts
both forms.

This issue is present with Keycloak, but other providers might be also
affected, as "scp" and "scope" seems to be the two common names for
returning scope claims.